+++ Classes/Domain/Repository/PageRepository.php
@@ -49,11 +49,11 @@ class PageRepository extends Repository implements RestRepositoryInterface
      *
      * @param int $offset
      * @param int $limit
-     * @param string $constraint
+     * @param array $constraints
      *
      * @return QueryResultInterface|array
      */
-    public function findByOffsetAndLimitAndConstraint($offset = 0, $limit = -1, $constraint = '')
+    public function findByOffsetAndLimitAndConstraints($offset = 0, $limit = -1, $constraints = [])
     {
         $query = $this->createQuery();
 
@@ -71,11 +71,11 @@ class PageRepository extends Repository implements RestRepositoryInterface
     /**
      * Count by constraint
      *
-     * @param string $constraint
+     * @param array $constraints
      *
      * @return int
      */
-    public function countByConstraint($constraint = '')
+    public function countByConstraints($constraints = [])
     {
         // we call countAll here, because we have no specific contraints
         return parent::countAll();